Threads now write to a shared output array. ... Which means that we do not have to allocate memory for separate output arrays, and then merge them once computations are done. This does however mean that the thread functions will have to be careful as to not write to the same memory locations.
Also added some different modes when distribution, which can be used to force the arrays to cover the same index ranges, or to make the distribution distribute and operate on the intersecting indices. There's also an option for executing parallel operations over the sum of the input arrays indices (which is useful for parallel concatenation, which I also implemented in this commit).